Don't store info about year in BaseEntry.
Deduce the year from the period instead. This fixes a bug that would
assign the current year when adding an entry to any period without
specifying a date.
Also a bug in creating recurrent entries is fixed. The year is not
displayed in the printout of the entries/models, hence most changes are
concerning string formatting.