Rethink error handling
On the level of the cli module, potential errors resulting from the
communication.run call are handled as fine-grained as possible.
Only in the worst case, the traceback of a generic Exception is printed.

This has several implications:
- the httprequests.CommunicationError and localserver.CommunicationError
classes are no longer aliases for Exception but derived exceptions
- httprequests._Proxy.run raises a CommunicationError if a bad response
is returned. This triggers offline storage of the original command line
request in cli (before, a dictionary containing an 'error' was returned
and handled in the communication module)
- httprequests._Proxy.run raises a ValueError if an unknown command is
given (unlikely since the command line parsing filters these out but who
knows how the package is used). 'stop' is an unknown command (hence
clean up in WebserviceTestCase.tearDown)
- A server-side error (in Server.run, and hence e.g. an error occurring
when dealing with a Period database) is returned as good response (if
communication not interrupted). The corresponding exception is raised in
communication.run. (This is still not ideal because any tracebacks from
generic server-side Exception are returned to the client and might
disclose internals)