Fix --filters option with flask backend
Two problems:
- the filters dict was not correctly sent along with the request
(inspected by enabling debugging of urllib3: b'filters=category'). The
print_parser in the resources module silently ignored this (and/or was not
even doing anything because nothing was to be parsed from the default
locations 'json' or 'values' of the request). Anyways, the unfiltered
database was returned
- as a fix, send the filters as comma-separated string and deserialize
it in the resources (which seems clunky)