Remove script for starting webservice
First, this was about not mixing client- and server-side config.
The config file is intended for client-side variables only.
The relevant variables are to be specified from the command line (which
supports the philosophy of having only one way to do things...)

It turns out that the flask cli already supports the exact same
behaviour with even more options. Hence the script is redundant and the
README is updated.