Change web host address format
This allows for direct communication with webservices using HTTPS.

On the client side, prefix with http(s)://. For starting the webservice,
do not prefix. Explanation:
requests expects a protocol like http:// as prefix, while flask will
fail to start an app using this address format because it splits it at
the colon to find out the port.