Change COPY_TAIL to /copy
Fixes #30.
It is more consistent to have the endpoint on 'first level' rather than after `/periods/copy`, since the CopyResource is a first-class resource like PeriodsResource. Also it avoids errors if anybody was to use a period named 'copy'.
However this is an incompatible change that introduces errors when financeager client and service of different versions (only relevant when using the Flask webservice).