Fix side-effects in filter test
self.pocket is set during the tests. If it is set to e.g. 1901, the
filter for the month '01' might accidentally find entries for the year
'01' (though that's not the target of the test). Hence set day and year
values to values that can't be confused with a YY-format year